feat: Update image loading and dependencies in about and index pages
The code changes include updating the image loading behavior in the about page and the index page. In the about page, the image class is modified to adjust the margin and the fetch priority is set to "high". In the index page, the script for changing the hero image on hover is removed and the hello-there image is commented out. Additionally, there are several dependency updates, including bumping astro-expressive-code to version 0.35.3, @astrojs/mdx to version 2.3.1, @astrojs/sitemap to version 3.1.4, eslint-plugin-markdown to version 5.0.0, eslint-plugin-astro to version 1.1.2, and astro to version 4.7.1.

Please note that the recent user commits and recent repository commits have been considered to generate this commit message.